+++ packages/node/tsup.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  dts: true,
+  sourcemap: true,
+  treeshake: true,
+  format: ["cjs", "esm"],
+});