+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tryabby/angular
 
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/core@5.1.3
+
 ## 2.0.4
 
 ### Patch Changes