+++ packages/next/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tryabby/next
 
+## 5.0.5
+
+### Patch Changes
+
+- @tryabby/react@5.0.5
+
 ## 5.0.4
 
 ### Patch Changes